Move carousel swiper and item into carousel component and extract icons
