+++ components/common/icon/Next.tsx
@@ -1,0 +1,17 @@
+function Next() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="32"
+      height="32"
+      viewBox="0 0 24 24"
+    >
+      <path
+        fill="currentColor"
+        d="M8.59 16.58L13.17 12L8.59 7.41L10 6l6 6l-6 6l-1.41-1.42Z"
+      />
+    </svg>
+  );
+}
+
+export default Next;
