+++ components/common/icon/Prev.tsx
@@ -1,0 +1,17 @@
+function Prev() {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="32"
+      height="32"
+      viewBox="0 0 24 24"
+    >
+      <path
+        fill="currentColor"
+        d="M15.41 16.58L10.83 12l4.58-4.59L14 6l-6 6l6 6l1.41-1.42Z"
+      />
+    </svg>
+  );
+}
+
+export default Prev;
